backend/foodgram_api/requirements.txt
asgiref==3.5.0
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.16
django-colorfield==0.4.2
django-filter==21.1
django-import-export==2.5.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.1
MarkupPy==1.14
MarkupSafe==2.1.1
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==9.1.0
pluggy==0.13.1
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.1.0
pyparsing==3.0.6
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tablib==3.2.0
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0
